Check schema in record batch stream
