+++ docs/python/api/scalars.rst
