+++ pyvortex/python/vortex/__init__.py
@@ -1,11 +1,14 @@
 from . import _lib
-from ._lib import expr, scalar
+from ._lib import arrays, encoding, expr, scalar
 from ._lib.dtype import DType, binary, bool_, float_, int_, null, struct, uint, utf8
-from .encoding import Array, array, compress
+from .arrays import Array, array
+from .compress import compress
 
 assert _lib, "Ensure we eagerly import the Vortex native library"
 
 __all__ = [
+    arrays,
+    encoding,
     expr,
     scalar,
     "Array",
