+++ pyvortex/python/vortex/_lib/compress.pyi
@@ -1,0 +1,3 @@
+import vortex as vx
+
+def compress(array: vx.Array) -> vx.Array: ...
