+++ pyvortex/python/vortex/_lib/encoding.pyi
@@ -1,38 +1,10 @@
 from typing import Any, final
 
-import numpy as np
-import pandas as pd
-import polars as pl
 import pyarrow as pa
 
 import vortex as vx
 
-@final
-class Array:
-    def __lt__(self, other: Array) -> Array: ...
-    def __le__(self, other: Array) -> Array: ...
-    def __gt__(self, other: Array) -> Array: ...
-    def __ge__(self, other: Array) -> Array: ...
-    def __len__(self) -> int: ...
-    def to_arrow_array(self) -> pa.Array: ...
-    @property
-    def encoding(self) -> str: ...
-    @property
-    def nbytes(self) -> int: ...
-    @property
-    def dtype(self) -> vx.DType: ...
-    def filter(self, mask: Array) -> Array: ...
-    def fill_forward(self) -> Array: ...
-    def scalar_at(self, index: int) -> Any: ...
-    def take(self, indices: Array) -> Array: ...
-    def slice(self, start: int, end: int) -> Array: ...
-    def tree_display(self) -> str: ...
-    def to_arrow_table(self) -> pa.Table: ...
-    def to_numpy(self, *, zero_copy_only: bool = True) -> np.ndarray: ...
-    def to_pandas_df(self) -> pd.DataFrame: ...
-    def to_polars_dataframe(self) -> pl.DataFrame: ...
-    def to_polars_series(self) -> pl.Series: ...
-    def to_pylist(self) -> list: ...
-
 def _encode(obj: Any) -> pa.Array: ...
-def compress(array: Array) -> Array: ...
+@final
+class BoolArray(vx.Array):
+    def __new__(cls, array: vx.Array) -> BoolArray: ...
