+++ pyvortex/python/vortex/_lib/mypy.allowlist
@@ -2,6 +2,8 @@
 
 # Not much point defining the exports or the modules themselves
 .*\.__all__
+vortex\._lib\.arrays
+vortex\._lib\.compress
 vortex\._lib\.dataset
 vortex\._lib\.dtype
 vortex\._lib\.encoding
