+++ pyvortex/python/vortex/compress.py
@@ -1,0 +1,3 @@
+from vortex._lib.compress import compress
+
+__all__ = ["compress"]
