+++ pyvortex/src/compress.rs
@@ -1,7 +1,18 @@
 use pyo3::prelude::*;
 use vortex::sampling_compressor::SamplingCompressor;
 
-use crate::encoding::PyArray;
+use crate::arrays::PyArray;
+use crate::install_module;
+
+pub(crate) fn init(py: Python, parent: &Bound<PyModule>) -> PyResult<()> {
+    let m = PyModule::new_bound(py, "compress")?;
+    parent.add_submodule(&m)?;
+    install_module("vortex._lib.compress", &m)?;
+
+    m.add_function(wrap_pyfunction!(compress, &m)?)?;
+
+    Ok(())
+}
 
 /// Attempt to compress a vortex array.
 ///
