+++ pyvortex/src/dtype/mod.rs
@@ -34,8 +34,8 @@ pub(crate) fn init(py: Python, parent: &Bound<PyModule>) -> PyResult<()> {
 }
 
 /// Base class for all Vortex data types.
-#[pyclass(name = "DType", module = "vortex")]
-#[derive(Clone)]
+#[pyclass(name = "DType", module = "vortex", frozen, eq, hash)]
+#[derive(Clone, PartialEq, Eq, Hash)]
 pub struct PyDType {
     inner: DType,
 }
